analytics-service/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
mysql-connector-python==8.0.26
requests==2.26.0
pymongo==3.12.0
numpy==1.21.2